requirements.txt
kafka-python>=2.0
yfinance>=0.2
snowflake-connector-python>=3.0
requests>=2.31
pandas>=2.2
plotly>=5.22
streamlit>=1.36
